requirements.txt
django
sqlalchemy
drf-spectacular
python-dotenv
mysqlclient
django-environ
djangorestframework